pyproject.toml
[tool.poetry]
name = "hex-machina-v2"
version = "0.1.0"
description = "AI-driven newsletter service that automatically monitors AI research and delivers concise summaries"
authors = ["Mathieu Crilout <mathieu.crilout@gmail.com>"]
readme = "README.md"
packages = [{include = "hex_machina", from = "src"}]
keywords = ["ai", "newsletter", "research", "automation", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
feedparser = "^6.0.10"
playwright = "^1.40.0"
playwright-stealth = "^1.0.5"
scrapy = "^2.11.0"
PyYAML = "^6.0.1"
MainContentExtractor = "*"
sqlalchemy = "^2.0.0"
duckdb = "^1.3.2"
duckdb-engine = "^0.17.0"
coverage = "^7.9.2"
scrapy-user-agents = "^0.1.1"
matplotlib = "^3.10.3"
pandas = "^2.3.1"
pathspec = "^0.12.1"
griffe = ">=0.20.0,<1.0.0"
markdown2 = "^2.5.3"
scrapy-playwright = "^0.0.43"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
ruff = "^0.1.6"
pre-commit = "^3.6.0"
bandit = "^1.8.6"
safety = "^3.6.0"
pytest-cov = "^6.2.1"
coverage = "^7.9.2"
scrapy-user-agents = "^0.1.1"
pytest-asyncio = "^1.1.0"





[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
target-version = ['py310']
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''



[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

 